Replace FATAL with ERROR for non-fatal failures
Whatever can be reasonably expected should not be FATAL. We should
use FATAL in only limited contexts, such as assertion failures or
system call errors that are never expected to fail.
FATAL incurs an abnormal exit, which is followed by core dump,
stack trace, and other consequences that we do not usually want.