Do not use exit() in a child process
gRPC crashes when a child process from frok() terminates with exit().
Some of its global (static) objects incorrectly free up resources that
are still being used by the parent process.

When VPort is created for a network namespace, the driver internally
uses fork() so that its child process can attach to the namespace and
set up IP addresses as specified. This patch replace exit() done in the
child process with _exit(), so that the offending object destructors can
be skipped.