Add support for function callstack in *.so files
While modules now can be loaded at runtime with shared objects, the
current function backtrace (crash log) does not correctly locate source
code from instruction pointers. This is because .so files can be
loaded at an arbitrary memory location; the base address of .so file
must be considered when corresponding memory addresses to position in
the original ELF file.