Merge LeafTrafficClass and Task.
Until now, a LeafTrafficClass could execute multiple Tasks in a
round-robin fashion.  This commit changes that: now a LeafTrafficClass
is always associated with exactly one Task.  This is done for multiple
reasons:

* We already have RoundRobinTrafficClass, we don't need to duplicate
  the round-robin logic in LeafTrafficClass.
* If each Task has its own TrafficClass, we can block them separately.
* To simplify the user interface.

The user (or controller) visible changes are:

* The AttachTask message is removed.
* Leaves created by the modules will be named with a '!leaf_' prefix.
  User created TC names cannot start with '!'.
* A new UpdateTcParent message is introduced: it can be used to move
  leaves in the tree (it is used by the attach_module python
  interface).
* If no parent is specified, a TC is orphan.  An orphan TC can be
  associated to a worker, otherwise it will be to one in a round-robin
  fashion.
* There's no default priority root traffic class, or default leaf.
  Instead, the scheduler is intitially empty. If a worker has a single
  orphan TC, it will become the root of the tree.  Otherwise all the
  orphans will be attached to a default round-robin TC.

To improve performance and make the code more generic, the Scheduler
and LeafTrafficClass now take the Task as a template parameter.  Only
Workers and Modules need to know the details about Task.  Once the
Scheduler reaches a leaf, the Task execute callback can be inlined.

As a result, scheduling a Leaf with a single Task appears now to be
slightly faster, according to traffic_class_bench.

On the other hand, if we want to run multiple tasks, we now have to
create a RoundRobinClass with multiple LeafTrafficClasses.  Compared
to a single LeafTrafficClass with multiple Tasks, this can be slower
(up to 5x for 256 tasks), because FinishAndAccountTowardsRoot() is called
after every Task execution.

Other changes in this commit:
* TrafficClass::Traverse() now provides the child arguments.  This is
  used to print "share" and "priority" in "show tc".
* WeightedFairTrafficClass and RoundRobinTrafficClass now contain an
  extra copy of pointers to the children.  They can be read from
  Traverse() while the workers are running.  This fixes a crash.