Merge LeafTrafficClass and Task.
Until now, a LeafTrafficClass could execute multiple Tasks in a
round-robin fashion.  This commit changes that: now a LeafTrafficClass
is always associated with exactly one Task.  This is done for multiple
reasons:

* We already have RoundRobinTrafficClass, we don't need to duplicate
  the round-robin logic in LeafTrafficClass.
* If each Task has its own TrafficClass, we can block them separately.
* To simplify the user interface.

The user (or controller) visible changes are:

* The AttachTask message is removed.
* The AddTC message now allows specifying a module name and a task id
  when a leaf is created.  It is used to attach tasks to the TC tree.
* Instead of creating a default Leaf for each worker, we now create a
  default RoundRobin for each worker.  The default parent for a new
  TC is now the default RoundRobin instead of the Root node.  When
  a module creates a new task, a Leaf will be attached to the default
  RoundRobin.
* By default, the Leaves created by the modules will be named with a
  '!' prefix.  User created TC names cannot start with '!'.  (This
  restriction could be removed if we allow Leaf nodes without names).

To improve performance and make the code more generic, the Scheduler
and LeafTrafficClass now take the Task as a template parameter.  Only
Workers and Modules need to know the details about Task.  Once the
Scheduler reaches a leaf, the Task execute callback can be inlined.

As a result, scheduling a Leaf with a single Task appears now to be
slightly faster, according to traffic_class_bench.

On the other hand, if we want to run multiple tasks, we now have to
create a RoundRobinClass with multiple LeafTrafficClasses.  Compared
to a single LeafTrafficClass with multiple Tasks, this can be slower
(up to 5x for 256 tasks), because FinishAndAccountTowardsRoot() is called
after every Task execution.