Fix sigsegv in cuckoo_map.h (#363)
Calling `*this = std::move(<something else>)` is unsafe outside of
the move constructor in C++11 (in particular *this is const). In this
case this resulted in a segfault. I changed this to a series of moves
of the inner elements. Interestingly in searching through github
(https://github.com/search?q="*this+%3D+std%3A%3Amove"+extension%3Acpp+extension%3Acxx&type=Code)
I have not found anyone use this outside of a move constructor, so I
wonder why we chose this construct.