Keep the ownership of FlowRecord within AvailablePorts
AvailablePorts used to give up the ownership of FlowRecord objects
after releasing a random pair of IP/port. This design is not safe,
as it depends on the caller to maintain the ownership correctly. When
AvailablePorts got destructed, released FlowRecord objects could
never be freed. Hence, a better design should keep the ownership
within AvailablePorts during its lifetime.