Embed leaf task v3 (#371)
* traffic_class_bench: Fix assertion.

TrafficClass::AddChild() expects the child to be constructed using
TrafficClassBuilder, otherwise it won't find it in the
TrafficClassBuilder::tc_count map.

The code is refactored in a future patch.  I wanted this fix on a
separate commit to run benchmarks before/after the refactor.

* traffic_class: Avoid caching number of descendants.

It's not read from the fast path.

This will make moving TCs around a little easier, since there's no need
to maintain a global map.

* traffic_class_bench: Use a dummy Module and Task.

A future commit will force LeafTrafficClass to have a Task (and a
Module).  This commit introduces it in the benchmark too, so that we can
make meaningful comparisons.

* traffic_class_bench: Single Leaf, many Tasks.

* traffic_class: Add RemoveChild() method.

This commit adds the RemoveChild() method to TrafficClass.  A new
BlockTowardsRoot() method is introduced to properly support removal.

It will be used in a future commit.

* Merge LeafTrafficClass and Task.

Until now, a LeafTrafficClass could execute multiple Tasks in a
round-robin fashion.  This commit changes that: now a LeafTrafficClass
is always associated with exactly one Task.  This is done for multiple
reasons:

* We already have RoundRobinTrafficClass, we don't need to duplicate
  the round-robin logic in LeafTrafficClass.
* If each Task has its own TrafficClass, we can block them separately.
* To simplify the user interface.

The user (or controller) visible changes are:

* The AttachTask message is removed.
* Leaves created by the modules will be named with a '!leaf_' prefix.
  User created TC names cannot start with '!'.
* A new UpdateTcParent message is introduced: it can be used to move
  leaves in the tree (it is used by the attach_module python
  interface).
* If no parent is specified, a TC is orphan.  An orphan TC can be
  associated to a worker, otherwise it will be to one in a round-robin
  fashion.
* There's no default priority root traffic class, or default leaf.
  Instead, the scheduler is intitially empty. If a worker has a single
  orphan TC, it will become the root of the tree.  Otherwise all the
  orphans will be attached to a default round-robin TC.

To improve performance and make the code more generic, the Scheduler
and LeafTrafficClass now take the Task as a template parameter.  Only
Workers and Modules need to know the details about Task.  Once the
Scheduler reaches a leaf, the Task execute callback can be inlined.

As a result, scheduling a Leaf with a single Task appears now to be
slightly faster, according to traffic_class_bench.

On the other hand, if we want to run multiple tasks, we now have to
create a RoundRobinClass with multiple LeafTrafficClasses.  Compared
to a single LeafTrafficClass with multiple Tasks, this can be slower
(up to 5x for 256 tasks), because FinishAndAccountTowardsRoot() is called
after every Task execution.

Other changes in this commit:
* TrafficClass::Traverse() now provides the child arguments.  This is
  used to print "share" and "priority" in "show tc".
* WeightedFairTrafficClass and RoundRobinTrafficClass now contain an
  extra copy of pointers to the children.  They can be read from
  Traverse() while the workers are running.  This fixes a crash.

* bessctl: Pretty print tree in 'show tc'.

* bessctl: Fix UpdateTc* messages

Since return_with_error() returns Status::OK, we need a different return
value for FindTC(), otherwise bessd will crash.

Also, there were a couple of problems with the update_tc_params python
function.

* traffic_class: Remove parent type from ChildArgs constructor.

The type can be inferred when constructing the parent.

* traffic_class: Use vector index instead of iterator.

Insted of doing arithmetic with iterators, it's simpler to use an
integer.

Also, there's no need to check 'next_index' for overflow, since we
already decrement it when necessary.

* Comment new functions and classes.

* Remove task.h and task-impl.h