Replace "ul" suffix with "ull" for 64-bit literals
On i686, integer literals XXXul can be considered 32-bit if XXX is
and small enough to fit in uint32_t. This may cause unexpected
behaviors, especially when bit shift operations are used.
Instead, use "ull" suffix when necesary, which is 64-bit on both
i686 and x86_64, regardless of the literal value of XXX.