travis: output coverage percentage from genhtml
Adding this output should make it clearer whether our fluctuating
code coverage is a problem with codecov.io or a problem with our
tests.

The coverage percentage generated by genhtml will not be the same as
the percentage generated by codecov.io (because aren't looking at
all of the same files and count differently for some line types),
but is roughly proportional.