bessctl.cc: Handle null gates without crashing.
The vector of gates in a module can legitimately contain null values.
We have to handle those without crashing (there was no crash in
{Enable,Disable}Tcpdump(), but we can improve error message).

This also fixes another bug in enable_track_for_module(): after enabling
tracking for a single gate, the function didn't return and it proceeded to
enable tracking for all gates (failing).

The commit also adds a few lines in exactmatch.bess as a regression test.