Detect scheduling constraint violations - Step 1
This change begins tracking placement constraints for tasks, e.g.,
recording when a physical port is best placed on a particular NUMA node.
We also add support for `bessctl` command `check violations` that can be
used to detect and find violations. This is the first step of this
change, and I am going to expand it to other port types, and also record
other types of violations.