bessctl: Allow CreateModule without stopping workers.
A new module can register tasks, but they will be initially orphans, so
they won't interfere with running workers.  It can register new metadata,
but the offset recomputation is postponed, so it won't interfere with
running workers.

The bessctl `add module` command now supports a `no_pause` option which
doesn't stop the workers.  When using this mode, new tasks created by the
module will be in an orphan state (not run by any worker).