bessctl: Allow ConnectModules without stopping workers.
If the modules are not used by any active workers, it is ok to connect
them without stopping workers.  Otherwise, pausing is still required,
because the gates are being changed, and the metadata offset need to be
recomputed.

The bessctl `add connection` command now supports a `no_pause` option
which doesn't stop the workers.