module: Optimize RunModuleSplit().
This commit adds a level of indirection to the splits_ thread local
variable: instead of keeping an entire batch for each possible output
gate, we keep a pointer to a batch.

This makes RunModuleSplit() faster, because we avoid copying the whole
batch (we only use a pointer now).

This also fixes Issue #443, because now we're using ~64K of stack space
instead of ~2M.

I tried running `samples/acl` on my laptop before and after the change
and looking at how many packets are generated by source0:

| source0 burst | before (Mpps) | after (Mpps) |
| -------------:| -------------:| ------------:|
|            32 |          76.0 |         83.2 |
|             1 |          14.3 |         14.9 |