Add and connect modules without stopping workers (#413)
* bessctl: Allow CreateModule without stopping workers.

A new module can register tasks, but they will be initially orphans, so
they won't interfere with running workers.  It can register new metadata,
but the offset recomputation is postponed, so it won't interfere with
running workers.

The bessctl `add module` command now supports a `no_pause` option which
doesn't stop the workers.  When using this mode, new tasks created by the
module will be in an orphan state (not run by any worker).

* bessctl: Extract collect_tc() function.

No functional change, it will be used by a future commit.

* bessctl: Include orphan TCs in ListTcs response.

It is useful to inspect the orphan TCs, since we're allowing module
creation without stopping the workers.

Sample output:

```
localhost:10514 $ show tc
<unattached>
  +-- !leaf_myq:0              leaf
<worker 0>
  +-- !default_rr_0            round_robin
      +-- !leaf_source0:0      leaf
      +-- !leaf_source1:0      leaf
```

* bessctl: Extract populate_active_worker().

The function will have another user in the next commit.  With this
change we iterate through the schedulers twice when checking the
constraints, but it's in not done in the fast path.

* bessctl: Allow ConnectModules without stopping workers.

If the modules are not used by any active workers, it is ok to connect
them without stopping workers.  Otherwise, pausing is still required,
because the gates are being changed, and the metadata offset need to be
recomputed.

The bessctl `add connection` command now supports a `no_pause` option
which doesn't stop the workers.

* bessctl: Remove unnecessary resume_all().

* bessctl: Specify default value in add command.