travis: avoid redundant builds for pull requests
Currently Travis launches two builds for pull requests submitted from
branches of NetSys/bess. This is the dominant workflow at the moment and
as a result Travis is doing a lot of extra work and the feedback
loop on pull requests is unnecessarily long.

This commit does away with the extra work by limiting Travis to only run
push jobs on the master branch.