add GateHookFactory
Currently, adding a GateHook requires adding an RPC which leads to lots
of extra code in bessctl.cc. This commit does away with this extra code
by moving it into the gate hooks themselves and adding GateHookFactory
to keep track of gate hooks and their initialization code similar to how
ModuleBuilder and PortBuilder keep track of Modules and Ports.