ip_lookup: work around DPDK out-of-memory bug
When we call rte_lpm_create() it returns NULL if the
parameters are invalid or if it runs out of memory.
For all but the "out of memory" case it sets rte_errno,
but when it runs out of memory, it fails to set rte_errno
at all.  This leaves whatever per-thread value it had
in it, which can often be zero.  This causes the subsequent
"return CommandFailure(rte_errno, ...)" to return a
failure with a zero error code, which pybess interprets
as success, so that it goes on to use the invalid data
and crashes bess.

While we probably should not crash, and perhaps should
have CommandFailure guarantee a nonzero return code or
otherwise arrange for failures to show up as failures,
we can work around this easily enough for now by pre-setting
ENOMEM as the error.