build.py: download dpdk via git
This enables simpler updating of BESS's DPDK dependency:
to experiment with a different release tag, or to use a
branch and maybe make your own changes, you can now enter
the deps/dpdk-17.05 directory, run "git fetch --unshallow",
and then use it as a normal Git repository.  The rest of
the build process will simply use the existing repository,
as long as it exists, thinking of it as "stock dpdk 17.05".

The unshallow step is needed only because we make a shallow
clone so that we download only about 14 MB (vs the old tarball
method which downloads about 12 MB; the full repository is
currently about 36 MB).

If you are building stock BESS with stock DPDK there is no
change at all in the process.