build.py: add verbose option
This can be used to observe the entire build, or for things
like Travis that like to kill builds if they remain quiet for
too long.

Note: we allow both V=1 in the environment (because we want to
pass that in to kmod builds), or a --verbose flag (because
arguments are convenient and easy in argparse).