rework module tests to avoid namespace issues
Regular *.bess files are run in their own private namespace,
which works well.

Test modules are not, and it makes writing tests trickier than
necessary.  Use the same technique that the standard file run
code uses, to provide the test namespace (which is noticeably
larger and includes three magic variables).

There is also now a place to put in a hook for optimization:
if we have not loaded or unloaded any plugins since we last
used gRPC to fetch all the BESS module-class and port-driver
names, they should still be the same.  (This may need a smaller
RPC, or a cache-flush command, if multiple bessctl instances
talk to a single daemon.)