remove 2to3 from protobuf generation
The main thing 2to3 was doing at this point was converting
"import ..." lines to read "from . import ...", i.e., making
everything a package-relative import, regardless of whether
things were intended to be package-relative.  This no longer
works if the plugins are generated into a separate plugin_pb
directory.

Removing 2to3 also speeds up the build a little bit.

On the negative side, under python3 (only), standalone unit
tests that do load protobuf code, but don't load bess.py,
need to add the builtin_pb directory to the path.  (Or,
if pybess/bess.py is loaded first, the test works anyway,
but let's not assume that the tests are imported in any
particular order.)