metadata: Fix crash when igate is nullptr.
`igate`s are stored in a vector, with the gate number used as index.
If a gate index in a module is not used, the vector will have nullptr
where the igate pointer should be.

Pipeline::TraverseUpstream() didn't check for nullptr.  This caused
a crash (reproducible with the test included in the commit).

The commit fixes the crash.