rework bess core Makefile for out-of-tree plugins
Use PLUGINS (which is to be set in core/extra.mk or similar)
to specify the paths of directories containing plug-in modules,
port drivers, protobufs, and/or utilities to build along with
the usual modules, drivers, protobufs, and/or utilities.

This allows experimental or otherwise not-part-of-core-BESS
code to be built without having to touch the core BESS repository
itself.

(A subsequent commit will make this usable from the top level
build.py script.)