container_build: mount plugin directories
The plugin source directories must be mounted inside the
container during a container build.

Note that the sample plugin is already mounted.  Similarly,
if two plugins are in /some/path and /some/path/sub/path,
we only need to mount the top one (and if they're listed in
backwards order we really should just mount the top one).