verify plugin top level directories exist
Before fixing container_build to pass plugin mount points
through, the container build silently discarded all the
plugins, since we deliberately allow a plugin to have only
a few items (e.g., only modules or only a port, for instance).
Add a dependency on the top level directories so that if they
are missing, the build errors-out.  This is not especially
pretty, but catches problems.