fix up coverage in travis
The coverage scan code was failing to apply to all the Python
tests (though clearly intended to do so) because bessctl uses
"sudo" to run bessd, and sudo clears out unknown environment
variables, including the ones needed to control gcov.

For now at least, use "sudo -E" to allow all environment
variables through.  This is not super-secure, but neither is
the rest of the code, so things are no worse than before.

Insert a short sleep in front of coverage, in case it takes
a little time to get the coverage data written.