Makefile: Use pkg-config if available.
Debian buster(testing) and presumably ubuntu 17.10 package all the
dependencies required to build BESS (except dpdk).

New versions of GRPC (packaged in debian) require libc-ares and if BESS
is statically linked (by default it is), `-lcares` must be passed to the
linker.

Instead of forcing a new dependency for all users (even the ones that don't
need it), this commit makes use of pkg-config, which will automatically
include the indirect dependencies (such as libc-ares).

For users that have an old grpc version (such as `container_build.py`),
pkg-config is still optional.  I'd like to make it required, but a new
container image must be generated, and that's a decision we can make
later.