ip_checksum: Handle VLAN and QinQ packets.
trafficgen udp mode now supports generating packet with a vlan tag.
This commit changes the ip_checksum module to correctly handle those
(and QinQ) packets.

As a result of this change, `ip_checksum` is now slower for untagged
packets.  In this very simple benchmark we measure the amount of traffic
generated by trafficgen udp mode with untagged packets before and after the
change on a single core:

BEFORE: 19.919 Mpps
AFTER: 19.444 Mpps

Also add a simple test for the ip_checksum module.