pybess: keep more names
We are adding get/set_runtime_config, etc., commands to (some)
BESS modules to make module save/restore easier.  Instead of
building up a module's internal state by:

 * creating the module
 * repeatedly "add"ing items

and only then allowing the module to run, complex modules like
ExactMatch can be created and then "set" once.

The existing commands all still work, so there is a corresponding
"get" to get the current effect of all additions and deletions,
or the most recently set state.  You can also "get_initial_arg"
to get the arguments that were, or could be, used to create the
module (it is not defined whether you get the original arguments
or the current values, if some of the arguments also show up in the
mutable state, although "current values" makes more sense in
general and would probably be recommended).

Note that get/set of runtime config can be thread-safe or
thread-unsafe, just like any other command.