bess: add state-save / restore for ExactMatch
The get_initial_arg command gets arguments that will
re-create the module.

The get_runtime_config and set_runtime_config commands get
and set the module's runtime configuration as a single entity,
rather than a series of commands that must be applied after
creating the module.

Add all three commands, and the ExactMatchConfig protobuf
that encodes the runtime configuration (the arguments are
naturally encoded via the existing ExactMatchArg).