fix race in unix_socket.cc
As unlikely as it seems, the Python tests for bessd are triggering
a race between the unix-socket accept() thread and the vlan test
code running in the Python interpreter.

There is a long comment in bessctl/test_utils.py explaining the race.
On the BESS side, we add a new boolean, `confirm_connect`, to the unix-
socket port, and send a magic number -- the byte-string "yes\0" --
across the socket to indicate that the accept thread has set it all up
for use.