bessctl/test_utils: fix single input packet
The test is currently coded to accept either a single
input packet -- if `input_pkts` is a `bytes` object or
something that can be converted to `bytes` -- or a list
of input packets (`if isinstance(input_pkts, list)`).

However, the loop that waits for packets to be processed
within bess simply uses `len(input_pkts)`.  So if you'
pass in a 60 byte packet, the loop waits for 60 packets
to be processed, rather than one packet.

As a simple fix, just convert non-list inputs to a list.