bessctl/test_utils: minor refactor/cleanup
Add an internal-use function that can be called at any point
to collect whatever output packets are available, without
otherwise disturbing the state of the test.  This is just a
minor refactoring I did while debugging the lost packets
in the unix_socket race, but I thought it worth committing.