worker_split: allow custom worker->gate mappings
It would be nice if WorkerSplit supported mappings other than f(x) = x.
This would be particularly useful in cases when BESS is configured with
a large number of workers and a module graph with a WorkerSplit module
that has only a few downstream modules. With the current implementation
this would require to a lot of edges from the WorkerSplit to the same
place.  For example, you might want to steer packets from workers on the
same CPU socket towrad the same destination.

To make these sorts of mappings a bit more concise, this commit changes
WorkerSplit to operate in one of two modes:

1) Packets from worker `x` are mapped to output gate `x`. This is the
   default mode.

2) When the `worker_gates` field is set, packets from a worker `x` are
   mapped to `worker_gates[x]`. In this mode, packet batches from
   workers not mapped to an output gate will be dropped.

The default mode(1) is about 2% slower (141Mpps vs 138Mpps) on my 2.1
GHz Xeon E5-2620 v4 after this change.