bess: add state-save / restore for WildcardMatch
Implement the same kind of module-argument and state get/set
handlers for WildcardMatch that are implemented for ExactMatch.

Provide a new test (using the new unittest framework) for
wildcard_match module, including the state save/restore.

Rip out the old WildcardMatchCommandGetRules.  It was never
converted to the new value_bin()-or-value_int() interface, so
it was unique in format and could not be used to restore the
module state.

It was also the only user of bin_to_uint64, so remove
that too.