url_filter: add state save/load, and make it work
Implements the same kind of module-argument and state
get/set handlers for url_filter that are implemented for
ExactMatch et al.

Provides a new test (using the new unittest framework) for url_filter
module, including the state save/restore.

Fixes the filter so that it passes the test (see below).

 ---

The existing URL filter didn't work at all any more: it got
broken when it was optimized to remove the separate "OK" and
"blocked" caches (commit c6376caf4405d5a94f086e535f5218e4d82d9597).
Fundamentally, we need a state flag bit in the flow cache entry,
since the packet sequence is:

   client -> server: SYN
   client <- server: SYN-ACK
   client -> server: ACK; data="GET <url> ..."

It's only at this point, when the <url> appears, that we
can decide whether to allow the flow to continue, or to block
it.  This means at least two packets must flow through the
filter in the client->server direction before we can make
a decision -- but the optimized flow cache contained only
"to block" entries, not "still need to analyze URL" entries.
Meanwhile the TCP reconstruction code must see the separate
SYN and data packets.

This would have blocked all flows, but there was a second bug:
flow cache entries, once added, had their expiry time set to
zero.  This meant that the second client-to-server packet
(containing the URL in its TCP data) looked up the flow entry
created by the SYN, immediately threw it away as expired, and
created a new flow entry.  The TCP reconstruction code then
immediately failed due to the lack of SYN, so we concluded
that the flow was OK and let the packet through.  The result
was to block *no* flows.

(I think the idea was to use "expiry time = 0" to mean "not
analyzed yet", but it didn't work as coded.  I find the new
clearer, and we can expire such entries later if/when BESS
gets a general timer system.)

As a smaller bug, we never set the IPv4 IDs in generated 403
and RST packets.  We now set them, although they're not really
right since we assume 1 and 2 are unique.  (The Python test
uses scapy's default IP ID of 1, so that the test passes.)

While here, fix up some documentation glitches too.