http_parser: fix alignment of ranges1 variable
The SSE based code makes use of some behavior not defined by
C compilers, but which will work on Intel hardware as long as
the variables are properly aligned so that they don't run off
the end of the address space.

To both guarantee operation and make clang's sanitizer happy,
fix the one missing alignment, and pad out the short arrays
so that they're 16 bytes long.  The latter messes slightly with
code clarity, so add a comment before findchar_fast.