url_filter: ensure compile-time packet sizes
When we generate our own packets, we copy them from a
template.  For the 403 response, we add data.  The data
themselves are constant, but the size wasn't.  We were doing
a variable size copy that read past the end of the for
speed.  Make the size constant, and let bess::utils::Copy()
default the last argument automatically.