rework control socket and gate-hook fifo handling
In existing BESS, only the control sockets use a thread,
which requires careful management of signals and system
calls.  We have some use cases for deferred open of FIFOs
that don't yet have readers (specifically, setting up
gate hooks where the order of operations is hard to control),
or even FIFOs where readers can disconnect and reconnect.

We move the signal code to new classes, with a new one for FIFO
operations.  This simplifies the control socket and existing
tcpdump and pcapng gate hooks.

There is a new test of the FIFOs that does at least a minimal
amount of testing against race conditions as well, which helps
verify that the control socket thread is not susceptible to
races.