minor module cleanups
The THREAD_{UN,}SAFE enumeration values are in the
Command namespace, not the Command::Command namespace.
This works due to some C++ quirk -- you can actually
write Command::Command::Command::Command::THREAD_SAFE,
for instance.  But it's misleading at best, so fix
two modules that use "overqualified" names.

In the code to check for thread-unsafe commands, we
do all the work to find out whether there are threads
running first, and then check whether that's OK.  Add
a mini-inline so that we can check, readably and yet
easily, in the other order.