bessctl.cc: simplify; fix missing delete
Add a small function to look up igate-or-ogate pointer.  Use
it in the function that actually creates gate hooks.

The section of code that created an input gate failed to
delete the allocated hook object in one particular error
case (hence leaking memory).  Both sub-classes of hooks
now share a single code path.