replace three-step hook creation with one-step
To create a new gate hook from its factory, you used to have
to do three calls in the correct order, with error checking:

  GateHook *hook = factory->CreateGateHook();
  CommandResponse init_ret = factory->InitGateHook(hook, gate, arg);
  // check init_ret for error here
  int add_ret = gate->AddHook(hook);
  // check add_ret for error here

One caller, the automatic Track hook, failed to check for errors
at all, and could (at least in theory) insert a tracking hook that
failed to initialize.  (The current Track init always succeeds,
though, and since the gate is new and therefore has no hooks yet,
the AddHook always succeeds as well.)

The other three just replicated the code.

Replace all this with:

  CommandResponse ret = gate->NewGateHook(factory, gate, is_igate, arg);

(where is_igate is a boolean telling us whether "gate" is an input
gate or output gate).  Now there is only one error to check (or
immediately return) and it's impossible to fail to delete the hook
instance if it did not initialize or get added.

To protect against accidental use of the old functions, delete
most of them.  The gate test code needs AddHook though, so move
it to private and give GateTest friend-access.