set hook->gate_ during hook construction
The existing code never connected hook instances to their gates.
Fortunately, no one was using this back-pointer; but we intend
to use it soon, so connect it up.

We could drop the "Gate *" argument to the hook instance
initializers (they can now use gate() to obtain it) but it's not
clear if changing this interface is a good idea, so we don't do
this here.