bessd_test: Use "fast" death style tests.
Death style tests (the ones that expect the process to die) are handled
by default using a very complicated procedure that involves fork-ing,
exec-ing(!), reexecuting the test from the beginning.

This interacts very poorly with our own DO_MULTI_PROCESS_TEST mechanism
because:
* It is a macro, so gtest has a hard time making sense of the line numbers
* We're forking on our own.

As a result, the tests spawn a lot of processes, some of them are blocked
and the testcase code is executed twice.

As a workaround, we can ask gtest to use the "fast" style of death tests,
which does not exec from the beginning.

Thanks a lot to @chris3torek for the debugging help and for finding a fix.