bessd_test: Use a separate tmp directory for each run.
If two users are concurrently executing the unit tests on the same machine,
they will interfere with each other because the tests in bessd_test use a
fixed path for the pidfile.

With this change we create a temporary directory for each run, to avoid
the problem.

Reported by @gatanu