Ensure the buffers argument is a power of 2
In accordance with rte_mempool_create() the number of buffers allocated should be
a power of 2 - 1. Command line option therefore forces the number to be a power of
two. packet.cc::init_mempool_socket() decrements the current try by 1.

Also check that the buffers argument is > minimum_try to be in accordance with
init_mempool_socket().