histogram: allow atomic and non-atomic usage
Allow atomic operations so that a histogram can be used
without a lock.

For backwards compatibility, the default is to use non-atomic
operations, requiring that the caller use a lock if the histogram
will be maniuplated in multiple threads.

Note that internally, this changes the counters from the
machine-dependent size_t to uint64_t (which affects 32-bit
implementations).