port: Return const reference from conf() accessor
According to Google C++ Style Guide we should return a const reference here
instead of making a copy.

I fixed this because I noticed that the accessor is used very often in the
fastpath ({Port,Queue}{Inc,Out}::RunTask call that once per batch). I haven't
noticed any performance impact, I suspect the compiler is smart enough to avoid
the copy, but I think it's better to change it.