build: Use a make jobserver if available
build.py launches make to build DPDK and BESS. If build.py is launched
from a parent make process, the -jN option (along a few others) given
to the parent would not propagate to the submake processes.

This commit checks if build.py script was invoked from parent make,
and if so, propagates make options to submake processes.