Allows specifying socket/numa node for BESS
Using "-n" flag a user may now specify the desired socket to allocate
huge pages from when using BESS.

Also includes a check for worker core existing on the correct socket
before trying to launch a worker. If worker is launched on a core on an
unused socket GetDefaultPool will crash bessd.

Signed-off-by: Tim Rozet <trozet@redhat.com>