vhost: fix the dataflow of vhost-user examples
The old behavior was that BESS generates packets and transmit to the
guest, then the guest forwards packets among ports, then BESS receives
and discards them.

With this PR, BESS does not actively create and feed packets to the
guest. Instead, both BESS and guest forwards packets, and the guest
starts the loop by creating some seed packets. This change was necessary
to prevent BESS overloading the VM or container even before they are
booting up.